requirements.txt
numpy>=1.21.0
requests>=2.25.0
matplotlib>=3.5.0
folium>=0.12.0
streamlit>=1.25.0
osmnx>=1.6.0
networkx>=3.1
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0 